인덱스 펀딩디테일 오류 수정, 펀딩 퍼센트 소수점 없앰(round 메소드 이용)

+++ src/main/java/creakok/com/domain/Funding.java
@@ -27,7 +27,7 @@ public class Funding {
 	private String funding_keyword;
 	private String funding_option;
 	private String funding_category_name;
-	private double percentage;
+	private int percentage;
 	private long restdays; 
 	private List<Funding> listrelatedFunding;
 	private Funding_qnaVo funding_qnaVo;
